feat: create standalone app
