+++ src/test/index.tsx
@@ -1,6 +1,6 @@
 import { prepareStore } from "@app/fx";
 import { Store, configureStore } from "@reduxjs/toolkit";
-import { waitFor } from "@testing-library/react";
+import { act, waitFor } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { RouteObject, RouterProvider, createMemoryRouter } from "react-router";
 import { REHYDRATE } from "redux-persist";
@@ -128,3 +128,13 @@ export const waitForToken = (store: Store<AppState>) => {
     }
   });
 };
+
+export const sleep = (n: number) =>
+  act(
+    () =>
+      new Promise<void>((resolve) => {
+        setTimeout(() => {
+          resolve();
+        }, n);
+      }),
+  );
