+++ src/ui/pages/create-project-git-settings.test.tsx
@@ -1,10 +1,14 @@
-import { CreateProjectGitSettingsPage } from "./create-project-git";
-import { defaultCodeScanResponse } from "@app/deploy/code-scan-result";
+import { fireEvent, render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { rest } from "msw";
+
+import { defaultCodeScanResponse } from "@app/deploy";
 import {
   createId,
   server,
   stacksWithResources,
   testApp,
+  testDatabasePostgres,
   testEnv,
 } from "@app/mocks";
 import {
@@ -12,8 +16,8 @@ import {
   createProjectGitSettingsUrl,
 } from "@app/routes";
 import { setupIntegrationTest, waitForToken } from "@app/test";
-import { fireEvent, render, screen } from "@testing-library/react";
-import { rest } from "msw";
+
+import { CreateProjectGitSettingsPage } from "./create-project-git-settings";
 
 describe("CreateProjectGitSettingsPage", () => {
   it("should render settings from url query param", async () => {
@@ -43,8 +47,34 @@ describe("CreateProjectGitSettingsPage", () => {
     await screen.findByText("postgres v14");
   });
 
-  it("should display error when validating environment variables", async () => {
+  it("should *not* render 'Connect Existing Database' button when no existing dbs exist", async () => {
     const handlers = stacksWithResources({ apps: [testApp] });
+    server.use(...handlers);
+
+    const { TestProvider, store } = setupIntegrationTest({
+      path: CREATE_PROJECT_GIT_SETTINGS_PATH,
+      initEntries: [createProjectGitSettingsUrl(`${testApp.id}`)],
+    });
+
+    await waitForToken(store);
+
+    render(
+      <TestProvider>
+        <CreateProjectGitSettingsPage />
+      </TestProvider>,
+    );
+
+    await screen.findByText(/Loading databases.../);
+    const btn = screen.queryByRole("button", {
+      name: /Connect Existing Database/,
+    });
+    expect(btn).not.toBeInTheDocument();
+  });
+
+  it("should display error when validating environment variables", async () => {
+    const handlers = stacksWithResources({
+      apps: [testApp],
+    });
     server.use(...handlers);
 
     const { TestProvider, store } = setupIntegrationTest({
@@ -79,7 +109,9 @@ describe("CreateProjectGitSettingsPage", () => {
   });
 
   it("should display error when validating database environment variables", async () => {
-    const handlers = stacksWithResources({ apps: [testApp] });
+    const handlers = stacksWithResources({
+      apps: [testApp],
+    });
     server.use(...handlers);
 
     const { TestProvider, store } = setupIntegrationTest({
@@ -103,6 +135,52 @@ describe("CreateProjectGitSettingsPage", () => {
     const envTxtArea = await screen.findByLabelText("Environment Variables");
     expect(envTxtArea.textContent).toBe("DEBUG=1\nAPP=");
     await screen.findAllByText("postgres v14");
+
+    const btn = await screen.findByRole("button", { name: "Save & Deploy" });
+    fireEvent.click(btn);
+
+    const text = await screen.findByText(
+      /DATABASE_URL has already been used, each database env var must be unique/,
+    );
+    expect(text).toBeInTheDocument();
+  });
+
+  it("should display error when validating database environment variables for new and existing dbs", async () => {
+    const handlers = stacksWithResources({
+      apps: [testApp],
+      databases: [testDatabasePostgres],
+    });
+    server.use(...handlers);
+
+    const { TestProvider, store } = setupIntegrationTest({
+      path: CREATE_PROJECT_GIT_SETTINGS_PATH,
+      initEntries: [
+        createProjectGitSettingsUrl(
+          `${testApp.id}`,
+          "dbs=database_url:postgres:14",
+        ),
+      ],
+    });
+
+    await waitForToken(store);
+
+    render(
+      <TestProvider>
+        <CreateProjectGitSettingsPage />
+      </TestProvider>,
+    );
+
+    await screen.findAllByText("postgres v14");
+
+    const existingDb = await screen.findByRole("button", {
+      name: /Connect Existing Database/,
+    });
+    fireEvent.click(existingDb);
+
+    const dbSelector = await screen.findByRole("combobox", {
+      name: /existing-db/,
+    });
+    userEvent.selectOptions(dbSelector, "test-app-1-postgres (postgres)");
 
     const btn = await screen.findByRole("button", { name: "Save & Deploy" });
     fireEvent.click(btn);
