+++ src/ui/shared/index.ts
@@ -49,3 +49,6 @@ export * from "./secret";
 export * from "./feedback";
 export * from "./footer";
 export * from "./stack";
+export * from "./progress-project";
+export * from "./status-box";
+export * from "./add-ssh-key";
